# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nimiq Mini Apps API",
    "version": "1.0.0",
    "description": "REST API for the [Nimiq Mini Apps](https://nimiqminiapps.com) community directory.\n\n**Submitting an app** — `POST /api/apps/submit` (wallet login required). Rate limit: **5 requests per hour per IP**.\nRequired field `submitter_contact` (Telegram, email, etc.) is stored privately and omitted from public responses.\n\n**Open in Nimiq Pay** — listed apps expose `open_url` as `https://nimpay.app/miniapps/open/{domain}` (domain without `https://`).\n\nMachine-readable spec: `GET /openapi.json` or `GET /openapi.yaml`.\n",
    "contact": {
      "name": "Nimiq Mini Apps",
      "url": "https://nimiqminiapps.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/NimMiniApps/NimiqMiniApps/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "https://api.nimiqminiapps.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/apps": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "List apps",
        "operationId": "listApps",
        "parameters": [
          {
            "$ref": "#/components/parameters/query"
          },
          {
            "$ref": "#/components/parameters/category"
          },
          {
            "$ref": "#/components/parameters/developer"
          },
          {
            "$ref": "#/components/parameters/tag"
          },
          {
            "$ref": "#/components/parameters/asset"
          },
          {
            "$ref": "#/components/parameters/rewards"
          },
          {
            "$ref": "#/components/parameters/status"
          },
          {
            "$ref": "#/components/parameters/featured"
          },
          {
            "$ref": "#/components/parameters/sort"
          },
          {
            "$ref": "#/components/parameters/collection"
          },
          {
            "$ref": "#/components/parameters/paginate"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/offset"
          }
        ],
        "responses": {
          "200": {
            "description": "App list, or paginated result when `paginate=1`",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AppPublic"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/PaginatedApps"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/apps/submit": {
      "post": {
        "tags": [
          "Submit"
        ],
        "summary": "Submit a new app for review",
        "description": "Creates an app with `status=submitted` and `featured=false`. Requires a wallet\nsession (see `POST /api/auth/verify`); the caller must have a `display_name` set\non their profile first. `developer_slug`/`developer_name` are derived from the\ncaller's wallet identity, not taken from the request body.\nRate limited to 5 submissions per hour per client IP (HTTP 429).\n",
        "operationId": "submitApp",
        "security": [
          {
            "walletCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppSubmitRequest"
              },
              "example": {
                "slug": "my-mini-app",
                "name": "My Mini App",
                "domain": "myapp.example.com",
                "category": "Games",
                "tagline": "One sentence pitch",
                "submitter_contact": "@telegram or you@example.com",
                "description": "Short plain-text summary for listings",
                "long_description": "## Features\n\n- **Bold** and lists work here",
                "tags": [
                  "games"
                ],
                "assets": [
                  "NIM"
                ],
                "reward_assets": [
                  "NIM"
                ],
                "release_stage": "beta"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "App submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppPublic"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "description": "Slug already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/apps/{slug}": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "Get app by slug",
        "operationId": "getApp",
        "parameters": [
          {
            "$ref": "#/components/parameters/slug"
          }
        ],
        "responses": {
          "200": {
            "description": "App details (submitted/rejected apps return 404 unless queried via admin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppPublic"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/apps/{slug}/status": {
      "get": {
        "tags": [
          "Submit"
        ],
        "summary": "Check submission review status",
        "operationId": "getSubmissionStatus",
        "parameters": [
          {
            "$ref": "#/components/parameters/slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Review status for a submitted or listed app",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionStatus"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/apps/{slug}/related": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "Related apps (same developer or category)",
        "operationId": "getRelatedApps",
        "parameters": [
          {
            "$ref": "#/components/parameters/slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Up to 4 related public apps",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppPublic"
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/apps/{slug}/track": {
      "post": {
        "tags": [
          "Catalog"
        ],
        "summary": "Track an open or view event (public beacon)",
        "description": "Fire-and-forget engagement counter. No auth. Rate limited to 20 requests/minute per IP+slug.\nUnknown slugs and malformed bodies return 204 (no user-facing error).\n",
        "operationId": "trackAppEvent",
        "parameters": [
          {
            "$ref": "#/components/parameters/slug"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "event"
                ],
                "properties": {
                  "event": {
                    "type": "string",
                    "enum": [
                      "open",
                      "view"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Event recorded or silently ignored"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/apps/{slug}/stats": {
      "get": {
        "tags": [
          "Submit"
        ],
        "summary": "Daily open/view stats for an app (owner or admin)",
        "operationId": "getAppStats",
        "security": [
          {
            "walletCookie": []
          },
          {
            "adminBearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Totals and last-30-days daily series (missing days omitted)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppStats"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Wallet session is not an owner or admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/apps/{slug}/request-update": {
      "post": {
        "tags": [
          "Submit"
        ],
        "summary": "Request a listing update (reviewed by admins)",
        "description": "Proposes changes to an existing public listing. Requires a wallet session for one\nof the app's current owners (see `owner_wallet_addresses`). `developer_name`/`developer_slug` in the\nrequest body are ignored — identity changes are admin-only. Rate limited (same\n5/hour per IP as submit). Only one pending revision per app at a time.\n",
        "operationId": "requestAppUpdate",
        "security": [
          {
            "walletCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/slug"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Update request queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateRequestCreated"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "The wallet session does not own this app",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "A pending update already exists for this app",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/apps/{slug}/owners": {
      "post": {
        "tags": [
          "Submit"
        ],
        "summary": "Add a co-owner wallet (self-service, caller must be a current owner)",
        "operationId": "addAppOwner",
        "security": [
          {
            "walletCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/slug"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "wallet_address"
                ],
                "properties": {
                  "wallet_address": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Owner added (or already present)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "added"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "The wallet session is not a current owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/apps/{slug}/owners/{wallet}": {
      "delete": {
        "tags": [
          "Submit"
        ],
        "summary": "Remove a co-owner wallet (self-service, caller must be a current owner)",
        "operationId": "removeAppOwner",
        "security": [
          {
            "walletCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/slug"
          },
          {
            "name": "wallet",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Owner removed"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "The wallet session is not a current owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Would remove the last remaining owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/my/apps": {
      "get": {
        "tags": [
          "Submit"
        ],
        "summary": "List apps owned by the logged-in wallet",
        "operationId": "myApps",
        "security": [
          {
            "walletCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Apps owned by the caller, each with has_pending_revision",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/AppAdmin"
                      },
                      {
                        "type": "object",
                        "properties": {
                          "has_pending_revision": {
                            "type": "boolean"
                          }
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/developers/{slug}": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "Get developer and their public apps",
        "operationId": "getDeveloper",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Developer profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Developer"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/admin/apps": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List all apps (any status)",
        "operationId": "adminListApps",
        "security": [
          {
            "adminBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Full catalog including submitted and rejected",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppAdmin"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Create app",
        "operationId": "adminCreateApp",
        "security": [
          {
            "adminBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "App created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppAdmin"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "description": "Slug already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Search users by display name or wallet address (for the developer picker)",
        "operationId": "adminSearchUsers",
        "security": [
          {
            "adminBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Prefix match on display_name or wallet_address; empty returns []",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "wallet_address": {
                        "type": "string"
                      },
                      "display_name": {
                        "type": "string",
                        "nullable": true
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/admin/apps/{slug}": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Replace app fields (merge semantics)",
        "operationId": "adminUpdateAppPut",
        "security": [
          {
            "adminBearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/slug"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "App updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppAdmin"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "Partially update app (merge semantics)",
        "operationId": "adminUpdateAppPatch",
        "security": [
          {
            "adminBearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/slug"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "App updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppAdmin"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete app",
        "operationId": "adminDeleteApp",
        "security": [
          {
            "adminBearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/slug"
          }
        ],
        "responses": {
          "204": {
            "description": "App deleted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/admin/apps/{slug}/owners": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Add a co-owner wallet (admin, no membership requirement)",
        "operationId": "adminAddAppOwner",
        "security": [
          {
            "adminBearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/slug"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "wallet_address"
                ],
                "properties": {
                  "wallet_address": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Owner added (or already present)"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/admin/stats": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Admin dashboard counts",
        "operationId": "adminStats",
        "security": [
          {
            "adminBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Pending submissions, unreachable domains, pending update requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminStats"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "adminBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Set ADMIN_TOKEN on the server; send as Authorization Bearer token"
      },
      "walletCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "wallet_session",
        "description": "Signed wallet-session cookie set by POST /api/auth/verify"
      }
    },
    "parameters": {
...
      "query": {
        "name": "q",
        "in": "query",
        "description": "Search name, tagline, description, tags, assets, reward assets, developer",
        "schema": {
          "type": "string"
        }
      },
...
    },
...
  }
}